Add test to check incorrect guesses array gets updated when the answer is inccorect and renamed teh sample data file name
